requirements.txt
streamlit
pandas
dnspython
python-whois
requests
openpyxl
tldextract
